Working encryption and decryption with static assignment
Algorithms all work when p and q are statically assigned unsigned integers.  They do not work if the mpz value was generated.  My guess is that the generated mpz values must be converted to unsigned int assignments somehow.

When static assignment is turned on, the code will verify algorithms work.